Put all the files into this project skeleton: |-- docs |-- scripts |-- src |    -- __init__.py |-- tests  -- .gitignore  -- environment.yml  -- README.md  -- setup.py
