+++ scripts/prep_for_vasp.py
